Fix Code and Get Invoice by SubOrderID

+++ src/main/java/com/smart/tailor/utils/response/BrandLaborQuantityResponse.java
@@ -18,13 +18,13 @@ public class BrandLaborQuantityResponse {
 
     private Integer laborQuantityMaxQuantity;
 
-    private Double laborQuantityMinPrice;
+    private Integer laborQuantityMinPrice;
 
-    private Double laborQuantityMaxPrice;
+    private Integer laborQuantityMaxPrice;
 
     private Boolean laborQuantityStatus;
 
-    private Double laborCostPerQuantity;
+    private Integer laborCostPerQuantity;
 
     private Boolean brandLaborQuantityStatus;
